Build:WIP:Working on object model.
