+++ src/i18n_tools/models/__init__.py
@@ -1,0 +1,2 @@
+from .corpus import Corpus, Message
+from .repository import Repository
