+++ src/i18n_tools/models/repository.py
@@ -1,0 +1,13 @@
+"""
+Repository module
+=================
+
+"""
+class Repository:
+    """
+    A class for handling i18n_tools repository structure.
+
+    This class provides functionality to build and organize an i18n_tools repository.
+    """
+    def __init__(self):
+        pass
